feat!: move ProviderScope and Preferences out of app_kit #245
