+++ lib/flutter_appkit.dart
@@ -2,8 +2,9 @@ export 'src/app.dart';
 export 'src/env.dart';
 export 'src/global_context.dart';
 export 'src/l10n/localization.dart';
-export 'src/locale.dart';
-export 'src/locale_notifier.dart';
 export 'src/logger.dart';
+export 'src/models/locale.dart';
 export 'src/net.dart';
 export 'src/preferences.dart';
+export 'src/providers/locale_notifier.dart';
+export 'src/widgets/locale_aware_widget.dart';
